requirement.txt
streamlit
pytest
python-dotenv
zxcvbn-python